Time: 3 ms (74.19%), Space: 69.7 MB (25.79%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,17 +1,17 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
         int n = nums.length;
-        double sum = 0;
+        int sum = 0;
         for(int i = 0 ; i < k ; i++){
             sum += nums[i];
         }
-        double max = sum;
+        int max = sum;
         for(int i = k ; i < n ; i++){
             sum += nums[i];
             sum -= nums[i - k];
             max = Math.max(max,sum);
         }
-        return max/k ;
+        return (double)max/k ;
     }
         
 }
